UT_Georeferencing_Alignment_Tin_12d_1
An example based on a TUM Ifc 4x1 file with georeferencing, alignments and tins.
Two Ifc 4x3 RC2 files are produced from the imported TUM data with the data in the Ifc 4x3 file in (a) Map coordinates  (b) Truncated coordinates with IfcMapConversions so that the final result is Map Coordinates.